Set build tags to methods DeleteNetworkInterface and DescribeVSwitchByID
