+++ pkg/aliyun/client/vpc.go
@@ -1,7 +1,6 @@
 package client
 
 import (
-	"context"
 	"fmt"
 	"time"
 
@@ -13,33 +12,6 @@ import (
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
 	"k8s.io/client-go/util/retry"
 )
-
-// DescribeVSwitchByID get vsw by id
-func (a *OpenAPI) DescribeVSwitchByID(ctx context.Context, vSwitchID string) (*vpc.VSwitch, error) {
-	req := vpc.CreateDescribeVSwitchesRequest()
-	req.VSwitchId = vSwitchID
-
-	l := log.WithFields(map[string]interface{}{
-		LogFieldAPI:       "DescribeVSwitches",
-		LogFieldVSwitchID: vSwitchID,
-	})
-
-	start := time.Now()
-	resp, err := a.ClientSet.VPC().DescribeVSwitches(req)
-	metric.OpenAPILatency.WithLabelValues("DescribeVSwitches", fmt.Sprint(err != nil)).Observe(metric.MsSince(start))
-	if err != nil {
-		l.WithField(LogFieldRequestID, apiErr.ErrRequestID(err)).Error(err)
-		return nil, err
-	}
-	l.WithField(LogFieldRequestID, resp.RequestId).Debugf("DescribeVSwitches: vsw slice = %+v, err = %v", resp.VSwitches.VSwitch, err)
-	if len(resp.VSwitches.VSwitch) == 0 {
-		return nil, apiErr.ErrNotFound
-	}
-	if len(resp.VSwitches.VSwitch) > 0 {
-		return &resp.VSwitches.VSwitch[0], nil
-	}
-	return nil, err
-}
 
 // AllocateEIPAddress create EIP
 func (a *OpenAPI) AllocateEIPAddress(bandwidth, chargeType, isp string) (*vpc.AllocateEipAddressResponse, error) {
